change default behavior to not fail
this updates the behavior of the "assert" action plugin such that
assert tasks are now nonfatal by default, and a new "fatal" keyword
will let you force an assertion to be fatal.

this means you can simply replace the default "assert" plugin with
this one in order to see the new behavior.